backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.1
yfinance==0.2.32
pandas==2.1.3
numpy==1.26.2
pydantic-settings==2.1.0
